pyproject.toml
[tool.poetry]
name = "python-learn-to-tracing"
package-mode = false
version = "0.1.0"
description = ""
authors = ["Artur Akmalov <artur@akmalov.com>"]
readme = "README.md"
packages = [{include = "python_learn_to_tracing"}]


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
loguru = "^0.7.2"
httpx = "^0.27.2"
opentelemetry-instrumentation-fastapi = "*"
opentelemetry-distro = "*"
opentelemetry-exporter-otlp = "*"
sqlalchemy = "*"
asyncpg = "*"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
alembic = "^1.13.3"
psycopg2-binary = "*"
opentelemetry-instrumentation-sqlalchemy = "*"
opentelemetry-instrumentation-asyncpg = "*"
redis = "*"
opentelemetry-instrumentation-redis = "^0.49b1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
